requirements.txt
flask==2.0.1
tensorflow==2.6.0
numpy==1.19.5
Pillow==8.3.1
werkzeug==2.0.1